Base Commit for Li-Ion Charger
This is the initial commit for the Li-Ion Charger DIY project.
It will provide the software for the automatic charging and discharging
of Li-Ion 18650 batteries. The aim of the software is to analyze already
used batteries in terms of the remaining capacity (mAh). The software
is built for n-battery setup and is switching between a charging and
discharging mode based on the current and voltage measured at the battery.

For the schematic of the PCB and the functionality, you will find explanations
in the source code as well as on my YouTube channel named HolzmannCFD and the
tag DIY.

Enjoy.