Added temperature feature (DS18B20) for a battery cell monitoring.
The user can specify a minimum and maximum temperature in which the pgrogramm (and cell analysis) should work. If for any reason, the cell temperature gets beyond the specified limits, the battery status is set to FAILED and the battery analysis can be modified to the users need. Right now, the program simply stops