Added LittleFS file system and corresponding classes
Added class WriterReader:
This class handles all IO from the battery. The battery can simply call all functions of the WriterReader class as it is added as a derivative to the Battery class.

Added class FileSystem:
This class handles all operation related to the LittleFS class and hence, simply write/read/append data to a given file. Furthermore, arbitrary file and folder manipulation are handled inside this class