Added storing of file data
Implemented the data writing of the charging and discharging process to a file named 'slot_<ID>'. After the battery is tested, the file name 'slot_<ID>' is renamed to the correct battery ID which one was tested (linear increment). Furthermore, the extension incorporated the starting point for adding final data to the written file such as:
   ++ Average capacity
   ++ Voltage after final charging (used for the 30-day analysis)
   ++ Time/Date (if needed)
   ++ Total number of discharging cycles
   ++ Standard deviation (if needed)
   ++ And others

During this development step, the classes were further modified and corrected